+++ __init__.py
@@ -2,4 +2,4 @@
 DBLift - A flexible database migration tool
 """
 
-__version__ = "2.4.2"
+__version__ = "2.6.0"